refactor(cli): split emit output helpers
